Split musicbrainz API types from client
The single musicbrainz.rs mixed the serde response shapes with the
request/caching client. Move the response DTOs into musicbrainz/types.rs
and keep the request functions, caches, MusicBrainzError, and
ReleaseSearchParams in musicbrainz/mod.rs, which re-exports the types so
crate::musicbrainz::* paths are unchanged. Inline tests move to
musicbrainz/tests.rs. Pure relocation; the internal response wrappers the
client reads gain pub(super) visibility to cross the new file boundary.
